Move reaper under shim package
